Mitigation: Use consistent (key) sizes and bounds parameters.
The implementation and the verification now use a 64-byte master key size
consistently. Additionally, the CTR_SIZE was also updated for HW-Mod-Auth.

Crucial memory addresses and size parameters were hardcoded throughout the code
base at various different places and in different languages (i.e., in Verilog
in the openMSP430 core, in Verilog in both versions of HW-Mod, in C in the
SW-Att wrapper, in LTL for both HW-Mod proof specs).

To hopefully help avoiding this kind of implementation-model mismatches in the
future, the entire code base was parameterized and can now be consistently
configured with cmake from a *single* place in the top-level CMakeLists.txt.